File Size 단위 수정 (Integer -> Double)
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@40034 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/display/epub/vo/EpubSeries.java
@@ -136,9 +136,9 @@ public class EpubSeries extends CommonInfo {
 	// Accrual
 	// ---------------------------------------------------
 	/** 참여자수 */
-	private Integer paticpersCnt;
+	private Double paticpersCnt;
 	/** 구매수 */
-	private Integer prchsCnt;
+	private Double prchsCnt;
 	/** 평균 평가 점수 */
 	private Double avgEvluScore;
 
@@ -328,7 +328,7 @@ public class EpubSeries extends CommonInfo {
 	 *
 	 * @return 참여자수
 	 */
-	public Integer getPaticpersCnt() {
+	public Double getPaticpersCnt() {
 		return this.paticpersCnt;
 	}
 
@@ -338,7 +338,7 @@ public class EpubSeries extends CommonInfo {
 	 * @param paticpersCnt
 	 *            참여자수
 	 */
-	public void setPaticpersCnt(Integer paticpersCnt) {
+	public void setPaticpersCnt(Double paticpersCnt) {
 		this.paticpersCnt = paticpersCnt;
 	}
 
@@ -347,7 +347,7 @@ public class EpubSeries extends CommonInfo {
 	 *
 	 * @return 구매수
 	 */
-	public Integer getPrchsCnt() {
+	public Double getPrchsCnt() {
 		return this.prchsCnt;
 	}
 
@@ -357,7 +357,7 @@ public class EpubSeries extends CommonInfo {
 	 * @param prchsCnt
 	 *            구매수
 	 */
-	public void setPrchsCnt(Integer prchsCnt) {
+	public void setPrchsCnt(Double prchsCnt) {
 		this.prchsCnt = prchsCnt;
 	}
